# Make mtz.object.delete_reflections() efficient