--- a/iotbx/mtz/object.h
+++ b/iotbx/mtz/object.h
@@ -577,11 +577,45 @@ namespace mtz {
        */
       void
       delete_reflections(af::const_ref<std::size_t> const& iref) {
+        /* only possible if reflections in memory */
+        IOTBX_ASSERT(ptr()->refs_in_memory);
+        int oldsize = ptr()->nref;
+        IOTBX_ASSERT(oldsize >= 0);
+        int newsize = oldsize - iref.size();
+        IOTBX_ASSERT(newsize >= 0);
         for (std::size_t i=iref.size(); i>0; i--) {
           // iref must be sorted in ascending order
           if (i > 1) IOTBX_ASSERT(iref[i-1] > iref[i-2]);
-          CMtz::MtzDeleteRefl(ptr(),iref[i-1]);
         }
+        for (int x = 0; x < ptr()->nxtal; ++x)
+          for (int s = 0; s < ptr()->xtal[x]->nset; ++s)
+            for (int c = 0; c < ptr()->xtal[x]->set[s]->ncol; ++c) {
+              // MTZCOL->ref are float*
+              float* newarray = NULL;
+              float* oldarray = ptr()->xtal[x]->set[s]->col[c]->ref;
+              ccp4array_new_size(newarray, newsize);
+
+              std::size_t skip_pointer = 0; // pointer into iref
+              std::size_t copy_pointer = 0; // pointer into newarray
+
+              for (std::size_t i = 0; i < oldsize; i++) {
+                // determine whether element i should be copied
+                if (skip_pointer < iref.size() && iref[skip_pointer] == i) {
+                  // skip i, advance skip_pointer
+                  skip_pointer++;
+                } else {
+                  // copy element i of oldarray into newarray at position copy_pointer
+                  newarray[copy_pointer++] = oldarray[i];
+                }
+              }
+              // Put new array in place
+              ptr()->xtal[x]->set[s]->col[c]->ref = newarray;
+              // Destroy old array
+              ccp4array_free(oldarray);
+              // Sanity check.
+              IOTBX_ASSERT(copy_pointer == newsize);
+            }
+        --ptr()->nref;
       }
 
     protected: