On deleting multiple entries create a new array and copy all remaining
elements into the new array in O(n).

This replaces the previous implementation where each element is removed
individually by shifting the remainder of the array one place forward
in O(n²).

Fixes #320