@graeme-winter 2019-09-18
Quick look, seems sensible - deleting reflections I have found problematic in the past so thanks for looking at this.