@rjgildea 2019-09-18
Tested this with the following script:
```
from __future__ import absolute_import, division, print_function

import time
import sys

import iotbx.mtz
from scitbx.array_family import flex

def run(args):
    assert len(args) == 1
    filename = args[0]
    obj = iotbx.mtz.object(filename)

    for n_delete in (1, 2, 5, 10, 20, 50, 100, 200, 500, 1000, 2000, 5000):
        nref_orig = obj.n_reflections()
        if n_delete > nref_orig:
            break
        iref = flex.sorted(flex.random_permutation(obj.n_reflections())[:n_delete])
        t0 = time.time()
        obj.delete_reflections(iref)
        t = time.time() -t0
        assert obj.n_reflections() == nref_orig - n_delete
        print("%4i: %fs" % (n_delete, t))


if __name__ == "__main__":
    run(sys.argv[1:])
```
on an example unmerged reflection file with >6 million observations `/dls/mx-scratch/rjgildea/performance_week/firehose/001/DataFiles/AUTOMATIC_DEFAULT_scaled_unmerged.mtz`.

Timings with original implementation:
```
   1: 0.038522s
   2: 0.074254s
   5: 0.080070s
  10: 0.268641s
  20: 0.449388s
  50: 0.898962s
 100: 2.601796s
 200: 4.928543s
 500: 12.040385s
1000: 24.746426s
2000: 49.419131s
5000: 119.079113s
```

Timings with new implementation:
```
   1: 0.161734s
   2: 0.171613s
   5: 0.153348s
  10: 0.150927s
  20: 0.149846s
  50: 0.153441s
 100: 0.151782s
 200: 0.155939s
 500: 0.149971s
1000: 0.156159s
2000: 0.149077s
5000: 0.156631s
```
The new implementation completes in a constant time regardless of the number of deleted reflections. It is marginally slower when deleting fewer than ~10 reflections, but significantly faster with many more reflections.