@Anthchirp 2019-09-18
slower for small `n` is to be expected as elements are copied individually rather than collectivelvia `memmove()` in a conditional loop, so unlikely to be optimized by the compiler. I'd say that a 0.1s cost is likely to be fine in all cases though and not worth the effort to optimize this further.