@Anthchirp 2019-09-19
I understood this as a programming exercise, and have changed the algorithm from an element-wise to a block-wise copy. The new implementation is now faster in all cases where N>1.

N | original | new
--- | --- | ---
   1 | 0.066203s | 0.271656s
   2 | 0.132401s | 0.071052s
   5 | 0.334103s | 0.044182s
  10 | 0.667886s | 0.045340s
  20 | 1.351719s | 0.045298s
  50 | 3.347832s | 0.045534s
 100 | 6.569747s | 0.044925s
 200 | 13.138928s  | 0.045242s
 500 | 32.789561s | 0.045194s
1000 | 66.570445s | 0.044691s
2000 | 131.479574s | 0.044265s
5000 | 329.646579s | 0.044940s

I have also validated the output by looking at the column values using:
```python
from __future__ import absolute_import, division, print_function                                                        
                                                                                                                        
import time                                                                                                             
import sys                                                                                                              
                                                                                                                        
import iotbx.mtz                                                                                                        
from scitbx.array_family import flex                                                                                    
                                                                                                                        
filename = "/dls/mx-scratch/rjgildea/performance_week/firehose/001/DataFiles/AUTOMATIC_DEFAULT_scaled_unmerged.mtz"     
if __name__ == "__main__":                                                                                              
    obj = iotbx.mtz.object(filename)                                                                                    
                                                                                                                        
    for n_delete in (1, 2, 5, 10, 20, 50, 100, 200, 500, 1000, 2000, 5000):                                             
        nref_orig = obj.n_reflections()                                                                                 
        if n_delete > nref_orig:                                                                                        
            break                                                                                                       
        permutation = flex.sorted(flex.random_permutation(obj.n_reflections()))                                         
        keep = permutation[n_delete:]                                                                                   
        iref = permutation[:n_delete]                                                                                   
        old_column = list(obj.columns())[0].extract_values()                                                            
        t0 = time.time()                                                                                                
        obj.delete_reflections(iref)                                                                                    
        t = time.time() -t0                                                                                             
        assert obj.n_reflections() == nref_orig - n_delete                                                              
        new_column = list(obj.columns())[0].extract_values()                                                            
        assert len(new_column) == len(old_column) - n_delete
        old_column = old_column.select(keep)                                                                            
        assert (old_column == new_column).all_eq(True)                                                                  
        print("%4i: %fs" % (n_delete, t))                           
```