@Anthchirp 2019-09-19
Question was asked why deleting 1 is slower than deleting 10. I've updated the code to delete (1, 1, 2, 1, 2, 5, 10, ...), which shows that this is initialisation overhead. As to why this happens in the new implementation but not in the old one I have no idea.